KeggContext in kegg.py -> metabolism/context.py
with corrected imports, of course
